LPS-146207 Split costumer portal components file to specific files
