+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/components/ProjectContacts/Skeleton.js
@@ -9,7 +9,7 @@
  * distribution rights of the Software.
  */
 
-import Skeleton from '../../../../common/components/Skeleton';
+import {Skeleton} from '../../../../common/components';
 
 const ProjectContactsSkeleton = () => {
 	return (
